requirements.txt
streamlit
mercadopago
google-generativeai
opencv-python-headless
